@@ -1,19 +1,12 @@
 package md.vnastasi.cloud.endpoint.model;
 
-import lombok.Builder;
-import lombok.NonNull;
-import lombok.Value;
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
-@Value
-@Builder
-public class NameHolder {
-
-    @NonNull
-    String shortName;
-
-    @NonNull
-    String middleName;
-
-    @NonNull
-    String longName;
+@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
+public record NameHolder(
+    @JsonProperty String shortName,
+    @JsonProperty String middleName,
+    @JsonProperty String longName
+) {
 }
